feat: DEV-2669: implement author filter search
