+++ src/common/Select/Select.styl
@@ -11,6 +11,7 @@
   &_size
     &_compact
       height 32px
+      border-radius 3px
 
     &_small
       height 24px
@@ -26,6 +27,7 @@
     display flex
     font-weight 500
     align-items center
+    height 100%
 
   &__value
     width max-content
