+++ src/tags/object/Paragraphs/Paragraphs.module.scss
@@ -69,6 +69,51 @@
   margin-left: 36px;
 }
 
+.collapsed {
+  display: block;
+  max-width: max-content;
+  height: 1px;
+  padding: 2px;
+  user-select: none;
+
+  & > * {
+    visibility: hidden;
+    height: 1px;
+    user-select: none;
+  }
+}
+
+.authorFilter {
+  width: 100%;
+  min-width: 320px;
+  max-width: max-content;
+  margin-bottom: 16px;
+
+  &__placeholder {
+    font-weight: 400;
+  }
+  &__search {
+    min-width: 318px;
+
+    &__input {
+      width: 100%;
+      height: 40px;
+      padding: 0 12px;
+      border: none;
+      border-bottom: 1px solid rgba(0, 0, 0, 0.15);
+      font-size: 14px;
+      line-height: 22px;
+      background: #fafafa;
+      outline: none;
+    }
+  }
+}
+
+.authorSearch {
+
+}
+
+
 .audio {
   height: 32px;
   width: 100%;
